refactor(cli): split source index show modules
